+++ NEWS.md
@@ -4,6 +4,10 @@
 
 - Improved how `node_color=` colours numeric node variables
 
+## Tests
+
+- Fixed the arc strength test depending on the BLAS
+
 # autograph 1.2.1
 
 ## Layouts